refactor(stm): moved PreparedIvcSnarkRequest to ivc prover_inputs
