+++ mithril-stm/src/proof_system/ivc_halo2_snark/mod.rs
@@ -8,4 +8,6 @@ pub(crate) mod verifier_setup;
 
 #[cfg(feature = "benchmark-internals")]
 pub(crate) use prover_input::IvcProverInput;
+pub(crate) use prover_input::PreparedIvcSnarkRequest;
+#[cfg(feature = "benchmark-internals")]
 pub(crate) use prover_setup::IvcSnarkProverSetup;
